metalit_balance/requirements.txt
asgiref==3.4.1
black==21.12b0
cffi==1.15.0
click==8.0.3
Django==4.0.1
djangorestframework==3.13.1
environs==9.4.0
marshmallow==3.14.1
mypy-extensions==0.4.3
mysqlclient==2.1.0
pathspec==0.9.0
platformdirs==2.4.1
pycparser==2.21
PyJWT==2.3.0
python-dotenv==0.19.2
pytz==2021.3
sqlparse==0.4.2
tomli==1.2.3
typing_extensions==4.0.1
tzdata==2021.5
